@@ -2,7 +2,7 @@
 //
 // This source file is part of the Hummingbird server framework project
 //
-// Copyright (c) 2021-2021 the Hummingbird authors
+// Copyright (c) 2021-2022 the Hummingbird authors
 // Licensed under Apache License v2.0
 //
 // See LICENSE.txt for license information
@@ -36,7 +36,7 @@ public struct HBRequest: HBSendableExtensible {
     /// Logger to use
     public var logger: Logger
     /// reference to application
-    public var application: HBApplication { self._internal.application }
+    public var application: HBApplication { self._internal.application.wrappedValue }
     /// Request extensions
     public var extensions: HBSendableExtensions<HBRequest>
     /// Request context (eventLoop, bytebuffer allocator and remote address)
@@ -61,8 +61,8 @@ public struct HBRequest: HBSendableExtensible {
 
     /// endpoint that services this request.
     internal var endpointPath: String? {
-        get { self._internal.endpointPath }
-        set { self._internal.endpointPath = newValue }
+        get { self._internal.endpointPath.wrappedValue }
+        set { self._internal.endpointPath.wrappedValue = newValue }
     }
 
     // MARK: Initialization
@@ -134,9 +134,9 @@ public struct HBRequest: HBSendableExtensible {
             self.version = version
             self.method = method
             self.headers = headers
-            self.application = application
+            self.application = .init(application)
             self.context = context
-            self.endpointPath = endpointPath
+            self.endpointPath = .init(endpointPath)
         }
 
         /// URI path
@@ -147,13 +147,14 @@ public struct HBRequest: HBSendableExtensible {
         let method: HTTPMethod
         /// Request HTTP headers
         let headers: HTTPHeaders
-        /// reference to application
-        let application: HBApplication
+        /// reference to application. Currently wrapped in HBUnsafeTransfer to make Sendable.
+        /// Hope to make it Sendable in the future
+        let application: HBUnsafeTransfer<HBApplication>
         /// request context
         let context: HBRequestContext
         /// Endpoint path. This is stored a var so it can be edited by the router. In theory this could
         /// be accessed on multiple thread/tasks at the same point but it is only ever edited by router
-        var endpointPath: String?
+        let endpointPath: HBUnsafeMutableTransferBox<String?>
     }
 
     private var _internal: _Internal
@@ -182,5 +183,5 @@ extension HBRequest: CustomStringConvertible {
 
 #if compiler(>=5.6)
 extension HBRequest: Sendable {}
-extension HBRequest._Internal: @unchecked Sendable {}
+extension HBRequest._Internal: Sendable {}
 #endif
